ai_service/requirements.txt
flask
sentence-transformers
Pillow
numpy
